Extract from full-node::run the CLI-related code
